+++ keyboards/handwired/polykybd/poly_util.h
@@ -4,9 +4,11 @@
 
 #include <stdint.h>
 
+
+void select_all_displays(void);
+
+void clear_all_displays(void);
+
 void display_message(uint8_t row, uint8_t col, const uint16_t* message, const GFXfont* font);
 
-// Clears all keycap displays then renders the "BOOT-LOADER!" message.
-// Called on master before reset_keyboard() and on slave from the
-// USER_SYNC_POLY_DATA handler when BOOTLOADER_DISPLAY is newly set.
 void display_bootloader_message(void);
